Copy type system specification from Xi
The morloc typesystem was previously being developed in the temporary Xi
repo. Since the typesystem is now ported into Morloc core, the
documentation should be as well.

The specification asciidoc does not integrate well with the rest of the
Morloc doc system since the specification is LaTeX based. Also, since
the specification is much more academic then most users would want, and
since it is likely to evolve into a paper, keeping it separate from the
rest of the documentation may be reasonable.