Commit code and begin a great culling
I'm going to start deleting everything that is either not good or not
accurate. If later on I want to get it back, it will be here in the git
graveyard.

I will begin the process of rewriting the entire technical document. I
will consider every single sentence and determine whether it contributes
to the narrative. If not, even if the content is good, I will cut it.