Add Waldmann2018 on why lists are bad
This will be important to the discussion of the importance of lists in
morloc. The paper argues that the Haskell standard list type is, being a
singly-linked list, limited in its use and should not have special
syntactic treatment. What is a morloc List? At minimum a morloc List is
an ordered set. This may be where I leave the definition. Functions such
as `length`, index access, or lookup by value will have different time
complexities in different languages. That is fine. Morloc will be
designed to be able to empirically discover the performance of a
function and use this knowledge to generate optimized programs.